+++ app/packages/core/src/components/Modal/Sidebar/Annotate/useCanAnnotate.ts
@@ -1,4 +1,3 @@
-import { FeatureFlag, useFeature } from "@fiftyone/feature-flags";
 import {
   canAnnotate,
   isGeneratedView,